+++ src/Engine/Graphics/Meshlet.h
@@ -7,10 +7,10 @@ namespace Seele
 struct Meshlet
 {
     AABB boundingBox;
-    uint16 uniqueVertices[Gfx::numVerticesPerMeshlet]; // unique vertiex indices in the vertex data
+    uint32 uniqueVertices[Gfx::numVerticesPerMeshlet]; // unique vertiex indices in the vertex data
     uint8 primitiveLayout[Gfx::numPrimitivesPerMeshlet * 3]; // indices into the uniqueVertices array, only uint8 needed
     uint32 numVertices;
     uint32 numPrimitives;
-    static void build(const Array<Vector>& positions, const Array<uint16>& indices, Array<Meshlet>& meshlets);
+    static void build(const Array<Vector>& positions, const Array<uint32>& indices, Array<Meshlet>& meshlets);
 };
 } // namespace Seele
